Rename queryservice package to query.service.
